Modificaciones

+++ src/main/java/uniandes/edu/co/proyecto/modelo/Bodega.java
@@ -13,14 +13,14 @@ public class Bodega {
     //Atributos
     @Id
     private String nombre;
-    private Float tamanio;
+    private Integer tamanio;
     
     @ManyToOne
     @JoinColumn(name = "nombre_sucursal", referencedColumnName = "nombre")
     private Sucursal nombre_sucursal;
 
     // Constructor
-    public Bodega(String nombre, Float tamanio, Sucursal nombre_sucursal) {
+    public Bodega(String nombre, Integer tamanio, Sucursal nombre_sucursal) {
         this.nombre = nombre;
         this.tamanio = tamanio;
         this.nombre_sucursal = nombre_sucursal;
@@ -34,7 +34,7 @@ public class Bodega {
         return nombre;
     }
 
-    public Float getTamanio() {
+    public Integer  getTamanio() {
         return tamanio;
     }
 
@@ -47,7 +47,7 @@ public class Bodega {
         this.nombre = nombre;
     }
 
-    public void setTamanio(Float tamanio) {
+    public void setTamanio(Integer  tamanio) {
         this.tamanio = tamanio;
     }
