+++ src/main/java/uniandes/edu/co/proyecto/repositorio/SucursalRepository.java
@@ -18,5 +18,5 @@ public interface SucursalRepository extends JpaRepository<Sucursal, String> {
     @Modifying
     @Transactional
     @Query(value = "INSERT INTO sucursales (nombre, tamanio, direccion, telefono, codigo_ciudad) VALUES (:nombre, :tamanio, :direccion, :telefono, :codigo_ciudad)", nativeQuery = true)
-    void insertSucursal(@Param("nombre") String nombre, @Param("tamanio") Float tamanio, @Param("direccion") String direccion, @Param("telefono") Integer telefono, @Param("codigo_ciudad") Integer codigo_ciudad);
+    void insertSucursal(@Param("nombre") String nombre, @Param("tamanio") Integer tamanio, @Param("direccion") String direccion, @Param("telefono") Integer telefono, @Param("codigo_ciudad") Integer codigo_ciudad);
 }
